requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.1.0
pydantic>=2.11.0
supabase>=2.15.0
sentence-transformers>=3.4.0
spacy>=3.8.7
langchain-groq>=0.3.0
groq>=0.23.0
datasets>=3.5.0
pandas>=2.2.0
